[transformation] Add partial_fit method to the pipeline.
In order to avoid fitting all the data in the concatenation of all
lives, the partial_fit method was added.

In this case, each life is iterated and the transformers are
partially fitted during the iteration. This procedure can
be seen in the fit method ofr the TransformerClass.

The LivesPipeline was added to allow this behaviour and each
transformer class added the partial_fit method